initial validation
