+++ src/main/kotlin/ak2/front/domain/service/Ak2UserService.kt
@@ -1,0 +1,14 @@
+package ak2.front.domain.service
+
+import ak2.front.domain.repositories.AkUserRepository
+import org.springframework.stereotype.Service
+
+@Service
+class Ak2UserService(
+  val akUserRepository: AkUserRepository
+) {
+
+  fun hasUserInfo(id: String): Boolean {
+    return akUserRepository.selectUserInfoById(id) != null
+  }
+}
